chore: InputBox 코드 이동
